add infrastructure build and deploy workflow
Why these changes are being introduced:

This commit adds the infrastructure deploy workflows to the this application,
asati. It adds new github automation, and updates the project's Makefile with
the new workflows.

A minor change was also made to the Dockerfile directing apt to perform a
`dist-upgrade` rather than an `upgrade`. A `dist-upgrade` will attempt to
perform all outstanding upgrades with deep dependency resolution and conflict
resolution, while an `upgrade` will simply perform outstanding upgrades.

How this addresses that need:

	* Add dev build workflow
	* Add stage build workflow
	* Add prod promotion workflow
	* Update Makefile w/ new workflows
	* Modify Dockerfile to do dist-upgrade

Side effects of this change:

None

Changes to be committed:
      new file:   .github/workflows/dev-build.yml
      new file:   .github/workflows/prod-promote.yml
      new file:   .github/workflows/stage-build.yml
      modified:   Dockerfile
      modified:   Makefile