stage container build on merge
Why these changes are being introduced:

This commit addresses the following issue:
The stage-build.yml is still configured to build and push an ECR image when a
PR is opened; it should push an ECR image when there is a push to main, as
would happen after the PR is approved and merged. This way, if more commits
are added during the PR review process, they will end up in the final ECR
image deployed to Stage. Currently stage-build is triggered in the same way
as dev-build meaning that any extra commits would not make it to the Stage
image.

How this addresses that need:

Modifies the build gh workflow to build the container on merge rather than
commit.

Side effects of this change:

None

Changes to be committed:
      modified:   .github/workflows/stage-build.yml